+++ src/views/pullrequest/pullRequestNode.ts
@@ -23,10 +23,14 @@ export interface FileDiffQueryParams {
 
 export interface PRFileDiffQueryParams extends FileDiffQueryParams {
     site: BitbucketSite;
+    repoHref: string;
     prHref: string;
     prId: number;
     participants: User[];
     commentThreads: Comment[][];
+    addedLines: number[];
+    deletedLines: number[];
+    lineContextMap: Object;
 }
 
 export class PullRequestTitlesNode extends AbstractBaseNode {